Actually remove global state. Also cleanup.
Mostly this involved removing (commenting out) some experimental
code used for playing with polling logic, with the eventual goal
of making an API full of non-blocking calls. For medical reasons,
I am unable to complete such designs right now, so I am abandoning
such attempts and relying on the original blocking logic.

The experimental code did a lot of no-nos like using global state,
duplicating code from other parts of the driver, and declaring
external linkages without the project's prefix (that is, non-static
functions that don't start with "ms8607_" in their names).
This wouldn't have been a problem if I were to have continued
working on that design and completed and polished it, but that
isn't what's happening. Commenting it out ensures that none of
those problems can affect the driver's usage as of this commit.

The `hsensor_read_user_register` function was added to the public
API as the `ms8607_hsensor_read_user_register` function because
I found it useful for testing the sensor's settings (even if it
isn't strictly necessary to use the sensor).