Forked MS8607 driver to MS5840.
(Actually, this isn't a fork, but it corresponds to a fork in the
original history of this repository's MS8607 and MS5840 drivers.
So instead of renaming files and symbols, we copy the files
(unlike the original fork), then we rename symbols and stuff
(like the original fork). The rest of the original commit message,
from the MS5840 repo, are below.)

At this point it's still just the MS8607 driver but with a
search-and-replace done to replace "ms8607" with "ms5840" and the README
updated to reflect the fact that this is intended for a different
driver. So it probably won't work for the ms5840 in this state, and
might not even compile. The more detailed work has yet to come.