MS5840: Adjusted math for MS5840 per its datasheet.
In previous commits, the values retrieved could be skewed or misleading
because this driver was using the MS8607's mathematics. This isn't an
oversight or anything; it just hadn't been changed, and I knew I'd need
to do it, and now I've done it. It is done. New math should be good.

At this point, the driver should be at the same level of competency as
the MS8607 driver that I've worked on. That is: It's not as good as I
want (it uses blocking calls and such), but it should at least yield
the most accurate/precise numbers that the sensor possibly can.