Bugfix. Add basic time-out logic to humidity acquisition.
This fixes a bug that would cause the driver to hang (infinite loop)
if the MS8607 behaved incorrectly.

To cause this behavior, the following things had to happen:
- Controller tells MS8607 to start computing a humidity reading
- MS8607 receives request
- Controller waits the prescribed amount of time
- Controller requests the result of the humidity reading
- MS8607 transmits NACK (Not Acknowledge bit)
- Controller waits the prescribed amount of time
- Controller requests the result of the humidity reading
- MS8607 transmits NACK (Not Acknowledge bit)
- Controller waits the prescribed amount of time
- Controller requests the result of the humidity reading
- MS8607 transmits NACK (Not Acknowledge bit)
-  ... etc ...

In other words, this only happens if the MS8607 keeps transmitting
NACK and never finishes computing the humidity reading.
This SHOULDN'T happen, but it did happen on my bench, presumably
with a damaged or defective unit.

It is, nonetheless, very important that the driver never
hangs (indefinite blocking) the client's code, even if the
MS8607 is not responding in reasonable ways.

This commit is a simple fix: if the driver waited 3x the prescribed
length of time for the reading to complete, and the sensor still
hasn't responded with a reading, it will give up and return an
error code indicating a NACK response.

(Ideally the driver API would have ways to do things in a completely
non-blocking fashion, but that would require rewriting most of the
driver, and I am just not up to it right now. It should be, for now,
Good Enough to simply place bounds on the amount of time that
a function may spend blocking.)

(Other side-note: once this commit was applied, the misbehaving
bench unit was able to eventually produce readings after it encountered
failures. Presumably it cleared up when the controller resent
the request to start computing a humidity reading, or perhaps
after reading pressure&temperature again. So it would still get
"stuck", but it could be "unstuck" by doing things besides
simply requesting the result over and over.)