Update README.md
This README update includes some verbage explaining that this
driver is not the original/official one, but instead has some
modifications (new/different features).

This commit also corresponds with a merger of the MS8607 and MS5840
driver repositories. This places the drivers for those sensors into
the same git repo, with the eventual goal of factoring common
components of those drivers into a single place (thus deduplicating
copypasta) and making a more uniform API that can handle
more than one sensor.