Big redo of error handling, start of unittests, etc.
This is also the beginning of merging some of the duplicated
code from the ms5840 and ms8607 drivers into the tepht_common.*
files.

Both of those drivers now share the same error codes, as well
as a more sophisticated error return type.

The error stringizing function has been replaced with an
error printing function. This is not as "simple", but it allows
error messages to be parameterized without introducing heap
allocation (so the drivers are still devoid of `malloc`).
The parameterization is important because it allows the
resulting error messages to, at the very least, reflect
which driver they came from, which is something that now
has to be solved with error handling code from the drivers
being merged together.