Bugfix + better testing for `tepht_int64_to_string`
This was a bug that might have never been encountered,
but whatever. The change involves ensuring that, if
the function is asked to print a number in an unusual base
(thus circumventing buffer size checks for common bases),
and the output is too large to fit in the provided buffer,
then it will truncate instead of overflowing the buffer.

A test was added for this.
Also added (somewhat unrelated) tests of the output length
return parameter.