Fix use-before-init bug that probably causes segfaults.
In an earlier commit I forgot to initialize the ms8607_sensor
driver's `context_accessor` fields inside the `ms8607_init_sensor`
function. When such an instance triggers errors that use this
context_accessor within `tepht_print_errors`, the uninitialized
pointers are dereferenced. This commit fixes the `ms8607_init_sensor`
function so that it populates all of the necessary fields.

Just to be safe, I checked the ms5840 driver, too.
It already initialized the context_accessor correctly.