Add min-width and padding to `tepht_(u)int64_to_string` function.
This also resulted in more `tepht_(u)int64_to_(binary|dec|hex)`
functions, with some having a `_default` suffix and lacking
min-width and padding parameters (they provide default values
for those parameters).

In addition, all of the prototypes for these functions are now
available in the header. I'm still not sure if they should
really be part of the public API, as they might not be there
forever, but given that the driver only has one user (probably)
right now, this is pretty low-risk. (And if the error reporting
code ever stops using them, we could still just, ya know,
leave them in.)